fixed train

+++ code/dataset.py
@@ -34,7 +34,7 @@ class SizeDataset(Dataset):
             self.transform = A.Compose([ 
                 A.ToFloat(max_value=255.0),             
                 ToTensorV2(),
-                torch.Tensor.double()
+                #torch.Tensor.double()
             ])  
         
         # index data into list
